requirements.txt
tortoise-orm
asyncpg
aerich
python-dotenv==0.13.0
aiogram==2.8
redis==3.5.0
aioredis==1.3.1
